Factor out repeat compilation check and put it in encoder
