+++ packages/codec/lib/index.ts
@@ -77,7 +77,12 @@ export {
   decodeReturndata,
   decodeRevert
 } from "./core";
-export { DecodingError, StopDecodingError, NoProjectInfoError } from "./errors";
+export {
+  DecodingError,
+  StopDecodingError,
+  NoProjectInfoError,
+  RepeatCompilationIdError
+} from "./errors";
 
 //now: what types should we export? (other than the ones from ./format)
 //public-facing types for the interface
